feat: Complete the list command operation
